refactor(BA-6671): move BaseFilterAdapter out of the data layer
BaseFilterAdapter (data/filter/adapter.py) was the sole data->models
back-edge in this area: it imports QueryCondition from models.clauses and
carries SQLAlchemy-typed conversion logic, violating the data-layer rules
(no SQLAlchemy, pure value objects only). It is not a data value type, and
all of its subclasses live in api/ and services/.

Move it to repositories/base/filter_adapter.py, which sits below both api
and services and above models, so every consumer resolves it in the correct
layer direction and its own models.clauses import is the correct
repositories->models direction. The QueryCondition/QueryOrder aliases stay
in models/clauses.py and common is untouched.

- Relocate adapter and remove the empty data/filter/ package
- Update all 20 importers (18 api, 1 service, 1 test) to the new path
