+++ src/ai/backend/manager/api/rest/auto_scaling_rule/adapter.py
@@ -21,11 +21,11 @@ from ai.backend.common.dto.manager.auto_scaling_rule import (
 from ai.backend.common.types import AutoScalingMetricSource
 from ai.backend.manager.data.deployment.scale_modifier import ModelDeploymentAutoScalingRuleModifier
 from ai.backend.manager.data.deployment.types import ModelDeploymentAutoScalingRuleData
-from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.endpoint.conditions import AutoScalingRuleConditions
 from ai.backend.manager.models.endpoint.orders import AutoScalingRuleOrders
 from ai.backend.manager.repositories.base import BatchQuerier, OffsetPagination
+from ai.backend.manager.repositories.base.filter_adapter import BaseFilterAdapter
 from ai.backend.manager.types import OptionalState, TriState
 
 __all__ = ("AutoScalingRuleAdapter",)
