+++ src/ai/backend/manager/api/rest/notification/adapter.py
@@ -30,7 +30,6 @@ from ai.backend.common.dto.manager.notification import (
     WebhookSpecResponse,
 )
 from ai.backend.common.dto.manager.notification.response import EmailSpecResponse
-from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.data.notification import (
     NotificationChannelData,
     NotificationRuleData,
@@ -47,6 +46,7 @@ from ai.backend.manager.models.notification.orders import (
     NotificationRuleOrders,
 )
 from ai.backend.manager.repositories.base import BatchQuerier, OffsetPagination
+from ai.backend.manager.repositories.base.filter_adapter import BaseFilterAdapter
 from ai.backend.manager.repositories.base.updater import Updater
 from ai.backend.manager.repositories.notification.updaters import (
     NotificationChannelUpdaterSpec,
