+++ src/ai/backend/manager/api/rest/resource_slot/adapter.py
@@ -15,12 +15,12 @@ from ai.backend.common.dto.manager.resource_slot.response import (
     NumberFormatDTO,
     ResourceSlotTypeDTO,
 )
-from ai.backend.manager.data.filter.adapter import BaseFilterAdapter
 from ai.backend.manager.data.resource_slot.types import ResourceSlotTypeData
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
 from ai.backend.manager.models.resource_slot.conditions import ResourceSlotTypeConditions
 from ai.backend.manager.models.resource_slot.orders import ResourceSlotTypeOrders
 from ai.backend.manager.repositories.base import BatchQuerier, OffsetPagination
+from ai.backend.manager.repositories.base.filter_adapter import BaseFilterAdapter
 
 __all__ = ("ResourceSlotAdapter",)
